backend/requirements.txt
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0 